edit note controllers

+++ src/controllers/index.note.controllers.js
@@ -8,7 +8,6 @@ crltNote.getNotes =async(req, res)=> {
 }
 
 crltNote.createNote= async(req, res)=>{
-
     const note = await new Note(req.body)
     note.userId = req.user.id
     await note.save()
@@ -24,18 +23,25 @@ crltNote.updateNoteById =async(req, res)=>{
     console.log(req.user)
     const id = req.params.id
     const note = await Note.findById(id)
-    console.log(note)
-    //const note = await Note.findByIdAndUpdate(id, req.body)
-   // console.log(note)
-    res.json('update note')
+    if(req.user.id == note.userId){
+        const noteUpdate = await Note.findByIdAndUpdate(id, req.body)
+       return res.json('update note')
+    }else {
+        res.json('note not found')
+    }
+  
 }
 
 crltNote.deleteNote =async(req, res)=>{
     const id = req.params.id
-    const note = await Note.findByIdAndDelete(id)
-    console.log(note)
-    res.json('delete note by Id')
+    const note = await Note.findById(id)
+    if(req.user.id == note.userId){
+        const noteDelete = await Note.findByIdAndDelete(id)
+       return res.json('delete note')
+    }else {
+        res.json('note not found')
+    }
+  
 }
-
 
 module.exports = crltNote
